requirements.txt
asgiref==3.4.0
certifi==2021.5.30
charset-normalizer==2.0.1
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==3.2
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==6.62.0
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
xlwt==1.3.0
